pyproject.toml
[project]
name = "iran-aqi-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "backoff>=2.2.1",
    "fastparquet>=2024.11.0",
    "httpx>=0.28.1",
    "jdatetime>=5.2.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "persiantools>=5.3.0",
    "pyarrow>=20.0.0",
    "python-bidi>=0.6.6",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
]
